core review, fixes, add more tests

+++ src/utils/FileSystem.cpp
@@ -1,50 +1,61 @@
 #include "FileSystem.h"
-#include "StringUtil.h"
+
+#include <dirent.h>
+#include <sys/stat.h>
+#include <unistd.h>
+
 #include <climits>
 #include <cstring>
 #include <ctime>
-#include <dirent.h>
 #include <iostream>
-#include <sys/stat.h>
-#include <unistd.h>
 #include <vector>
+
+#include "StringUtil.h"
 
 #define MAX_EXT_LENGTH 8
 
 using namespace WebSocketCpp;
 
-std::string FileSystem::GetFullPath(const std::string &path) {
-    char buffer[PATH_MAX];
-    char *retval = realpath(path.c_str(), buffer);
+std::string FileSystem::GetFullPath(const std::string& path)
+{
+    char  buffer[PATH_MAX];
+    char* retval = realpath(path.c_str(), buffer);
 
-    if (retval == nullptr) {
+    if (retval == nullptr)
+    {
         return "";
     }
 
     return std::string(retval);
 }
 
-std::string FileSystem::GetApplicationFolder() {
+std::string FileSystem::GetApplicationFolder()
+{
     char buffer[PATH_MAX];
-    int len = readlink("/proc/self/exe", buffer, 250);
-    if (len >= 0) {
+    int  len = readlink("/proc/self/exe", buffer, 250);
+    if (len >= 0)
+    {
         return std::string(buffer, len);
     }
 
     return "";
 }
 
-bool FileSystem::ChangeDir(const std::string &path) {
+bool FileSystem::ChangeDir(const std::string& path)
+{
     return (chdir(path.c_str()) == 0);
 }
 
-std::string FileSystem::NormalizePath(const std::string &path, bool file) {
+std::string FileSystem::NormalizePath(const std::string& path, bool file)
+{
     std::string tmp = path;
     StringUtil::Replace(tmp, "~/", HomeFolder() + "/");
     StringUtil::Replace(tmp, "//", "/");
 
-    if (file == false) {
-        if (tmp.rfind(FileSystem::PathDelimiter()) != (tmp.length() - 1)) {
+    if (file == false)
+    {
+        if (tmp.rfind(FileSystem::PathDelimiter()) != (tmp.length() - 1))
+        {
             return tmp + FileSystem::PathDelimiter();
         }
     }
@@ -52,52 +63,66 @@ std::string FileSystem::NormalizePath(const std::string &path, bool file) {
     return tmp;
 }
 
-std::string FileSystem::ExtractFileName(const std::string &path) {
+std::string FileSystem::ExtractFileName(const std::string& path)
+{
     size_t i = path.rfind(FileSystem::PathDelimiter(), path.length());
-    if (i != std::string::npos) {
+    if (i != std::string::npos)
+    {
         return (path.substr(i + 1, path.length() - i));
     }
 
     return path;
 }
 
-std::string FileSystem::ExtractFileExtension(const std::string &path) {
+std::string FileSystem::ExtractFileExtension(const std::string& path)
+{
     size_t i = path.rfind('.', path.length());
-    if (i != std::string::npos && (path.length() - i) < MAX_EXT_LENGTH) {
+    if (i != std::string::npos && (path.length() - i) < MAX_EXT_LENGTH)
+    {
         return (path.substr(i + 1, path.length() - i));
     }
 
     return "";
 }
 
-bool FileSystem::IsFileExist(const std::string &path) {
-    bool retval = false;
+bool FileSystem::IsFileExist(const std::string& path)
+{
+    bool        retval = false;
     struct stat sb;
 
-    try {
-        if (stat(path.c_str(), &sb) == 0) {
+    try
+    {
+        if (stat(path.c_str(), &sb) == 0)
+        {
             retval = true;
         }
-    } catch (std::exception &ex) {
+    }
+    catch (std::exception& ex)
+    {
         std::cout << "reading file attributes failed: " << ex.what() << std::endl;
     }
     return retval;
 }
 
-int FileSystem::GetFileSize(const std::string &path) {
-    int fileSize = -1;
+int FileSystem::GetFileSize(const std::string& path)
+{
+    int         fileSize = -1;
     struct stat sb;
-    try {
+    try
+    {
         if (stat(path.c_str(), &sb) == 0)
             fileSize = sb.st_size;
-    } catch (std::exception &ex) {
+    }
+    catch (std::exception& ex)
+    {
         std::cout << "reading file size failed: " << ex.what() << std::endl;
     }
 
     return fileSize;
 }
 
-char FileSystem::PathDelimiter() {
+char FileSystem::PathDelimiter()
+{
 #ifdef _WIN32
     return '\\';
 #else
@@ -105,14 +130,17 @@ char FileSystem::PathDelimiter() {
 #endif
 }
 
-std::string FileSystem::Root() {
+std::string FileSystem::Root()
+{
 #ifdef _WIN32
-    char const *folder = getenv("SystemDrive");
-    if (folder == nullptr) {
+    char const* folder = getenv("SystemDrive");
+    if (folder == nullptr)
+    {
         return "\\";
     }
     std::string root(folder);
-    if (root.at(root.size() - 1) != FileSystem::PathDelimiter()) {
+    if (root.at(root.size() - 1) != FileSystem::PathDelimiter())
+    {
         root += FileSystem::PathDelimiter();
     }
     return root;
@@ -121,10 +149,13 @@ std::string FileSystem::Root() {
 #endif
 }
 
-bool FileSystem::IsDir(const std::string &path) {
+bool FileSystem::IsDir(const std::string& path)
+{
     struct stat result;
-    if (stat(path.c_str(), &result) == 0) {
-        if ((result.st_mode & S_IFDIR) != 0) {
+    if (stat(path.c_str(), &result) == 0)
+    {
+        if ((result.st_mode & S_IFDIR) != 0)
+        {
             return true;
         }
     }
@@ -132,25 +163,28 @@ bool FileSystem::IsDir(const std::string &path) {
     return false;
 }
 
-std::vector<FileSystem::FileInfo> FileSystem::GetFolder(const std::string &path)
+std::vector<FileSystem::FileInfo> FileSystem::GetFolder(const std::string& path)
 {
     std::vector<FileSystem::FileInfo> retval;
 
-    DIR *dir;
-    struct dirent *diread;
-    struct stat info;
+    DIR*           dir;
+    struct dirent* diread;
+    struct stat    info;
 
-    if ((dir = opendir(path.c_str())) != nullptr) {
-        while ((diread = readdir(dir)) != nullptr) {
-            long size = (-1);
+    if ((dir = opendir(path.c_str())) != nullptr)
+    {
+        while ((diread = readdir(dir)) != nullptr)
+        {
+            long        size = (-1);
             std::string mod;
-            if (stat(std::string(NormalizePath(path) + diread->d_name).c_str(), &info) == 0) {
-                auto mod_time = info.st_mtime;
-                const time_t *t = &mod_time;
-                struct tm *timeinfo = gmtime(t);
-                char buffer[30];
+            if (stat(std::string(NormalizePath(path) + diread->d_name).c_str(), &info) == 0)
+            {
+                auto          mod_time = info.st_mtime;
+                const time_t* t        = &mod_time;
+                struct tm*    timeinfo = gmtime(t);
+                char          buffer[30];
                 strftime(buffer, 30, "%a, %d %b %Y %H:%M:%S GMT", timeinfo);
-                mod = buffer;
+                mod  = buffer;
                 size = info.st_size;
             }
 
@@ -162,28 +196,36 @@ std::vector<FileSystem::FileInfo> FileSystem::GetFolder(const std::string &path)
     return retval;
 }
 
-bool FileSystem::CreateFolder(const std::string &path) {
+bool FileSystem::CreateFolder(const std::string& path)
+{
     struct stat s;
     std::string fullpath = "";
-    bool retval = false;
+    bool        retval   = false;
 
-    try {
+    try
+    {
         auto list = StringUtil::Split(path, FileSystem::PathDelimiter());
-        for (auto &folder : list) {
-            if (fullpath.empty() && folder.empty()) {
+        for (auto& folder : list)
+        {
+            if (fullpath.empty() && folder.empty())
+            {
                 fullpath = FileSystem::Root();
                 continue;
             }
             fullpath += ((fullpath.empty() || fullpath == FileSystem::Root())
-                             ? ""
-                             : std::string(1, FileSystem::PathDelimiter()))
-                        + folder;
-            if (stat(fullpath.c_str(), &s) == 0) {
-                if (S_ISDIR(s.st_mode)) {
+                                ? ""
+                                : std::string(1, FileSystem::PathDelimiter())) +
+                        folder;
+            if (stat(fullpath.c_str(), &s) == 0)
+            {
+                if (S_ISDIR(s.st_mode))
+                {
                     retval = true;
                     continue;
                 }
-            } else {
+            }
+            else
+            {
 #if defined __linux__ || defined __APPLE__
                 if (mkdir(fullpath.c_str(), S_IRWXU | S_IRWXG | S_IROTH | S_IXOTH) != 0)
 #else
@@ -191,39 +233,55 @@ bool FileSystem::CreateFolder(const std::string &path) {
 #endif
                 {
                     retval = false;
-                } else {
+                }
+                else
+                {
                     retval = true;
                 }
             }
         }
 
         return retval;
-    } catch (std::exception &ex) {
+    }
+    catch (std::exception& ex)
+    {
         std::cout << "exception while creating a folder: " << ex.what() << std::endl;
         return false;
     }
 }
 
-bool FileSystem::DeleteFolder(const std::string &path) {
-    DIR *dir;
-    struct stat s;
-    struct dirent *entry;
+bool FileSystem::DeleteFolder(const std::string& path)
+{
+    DIR*           dir;
+    struct stat    s;
+    struct dirent* entry;
 
-    try {
-        if (stat(path.c_str(), &s) == 0) {
+    try
+    {
+        if (stat(path.c_str(), &s) == 0)
+        {
             dir = opendir(path.c_str());
-            while ((entry = readdir(dir))) {
-                std::string file(entry->d_name);
-                if (file == "." || file == "..") {
-                    continue;
-                }
+            if (dir)
+            {
+                while ((entry = readdir(dir)))
+                {
+                    std::string file(entry->d_name);
+                    if (file == "." || file == "..")
+                    {
+                        continue;
+                    }
 
-                file = FileSystem::NormalizePath(path) + file;
-                if (stat(file.c_str(), &s) == 0) {
-                    if (S_ISDIR(s.st_mode)) {
-                        DeleteFolder(file);
-                    } else {
-                        unlink(file.c_str());
+                    file = FileSystem::NormalizePath(path) + file;
+                    if (stat(file.c_str(), &s) == 0)
+                    {
+                        if (S_ISDIR(s.st_mode))
+                        {
+                            DeleteFolder(file);
+                        }
+                        else
+                        {
+                            unlink(file.c_str());
+                        }
                     }
                 }
             }
@@ -231,34 +289,41 @@ bool FileSystem::DeleteFolder(const std::string &path) {
             rmdir(path.c_str());
 
             return true;
-        } else {
+        }
+        else
+        {
             throw "cannot access the folder";
         }
-    } catch (std::exception &ex) {
+    }
+    catch (std::exception& ex)
+    {
         std::cout << "exception while removing the folder: " << ex.what() << std::endl;
         return false;
     }
 }
 
-std::string FileSystem::GetDateTime() {
-    time_t tm;
-    struct tm *timeinfo;
-    char buffer[30];
+std::string FileSystem::GetDateTime()
+{
+    time_t     tm;
+    struct tm* timeinfo;
+    char       buffer[30];
 
     time(&tm);
-    const time_t *t = &tm;
-    timeinfo = gmtime(t);
+    const time_t* t = &tm;
+    timeinfo        = gmtime(t);
     strftime(buffer, 30, "%a, %d %b %Y %H:%M:%S GMT", timeinfo);
     return buffer;
 }
 
-std::string FileSystem::GetFileModifiedTime(const std::string &file) {
+std::string FileSystem::GetFileModifiedTime(const std::string& file)
+{
     struct stat result;
-    if (stat(file.c_str(), &result) == 0) {
-        auto mod_time = result.st_mtime;
-        const time_t *t = &mod_time;
-        struct tm *timeinfo = gmtime(t);
-        char buffer[30];
+    if (stat(file.c_str(), &result) == 0)
+    {
+        auto          mod_time = result.st_mtime;
+        const time_t* t        = &mod_time;
+        struct tm*    timeinfo = gmtime(t);
+        char          buffer[30];
         strftime(buffer, 30, "%a, %d %b %Y %H:%M:%S GMT", timeinfo);
         return buffer;
     }
@@ -266,35 +331,45 @@ std::string FileSystem::GetFileModifiedTime(const std::string &file) {
     return "";
 }
 
-void RandInit() { srand(static_cast<unsigned int>(time(nullptr))); }
+void RandInit()
+{
+    srand(static_cast<unsigned int>(time(nullptr)));
+}
 
-uint32_t GetRand(uint32_t min, uint32_t max) {
+uint32_t GetRand(uint32_t min, uint32_t max)
+{
     return static_cast<uint32_t>(rand()) % (max - min) + min;
 }
 
-std::string FileSystem::TempFolder() {
+std::string FileSystem::TempFolder()
+{
     std::vector<std::string> tmpFolders = {"TMPDIR", "TEMP", "TMP", "TEMPDIR"};
 
-    char const *folder = nullptr;
-    for (auto &env : tmpFolders) {
+    char const* folder = nullptr;
+    for (auto& env : tmpFolders)
+    {
         folder = getenv(env.c_str());
-        if (folder == nullptr) {
+        if (folder == nullptr)
+        {
             continue;
         }
     }
 
-    if (folder == nullptr) {
+    if (folder == nullptr)
+    {
         folder = "/tmp";
     }
 
     std::string path(folder);
 
-    if (path.at(path.size() - 1) != FileSystem::PathDelimiter()) {
+    if (path.at(path.size() - 1) != FileSystem::PathDelimiter())
+    {
         path += FileSystem::PathDelimiter();
     }
 
     RandInit();
-    for (int i = 0; i < 6; i++) {
+    for (int i = 0; i < 6; i++)
+    {
         char ch = GetRand(static_cast<int>('a'), static_cast<int>('z'));
         path.push_back(ch);
     }
@@ -302,4 +377,8 @@ std::string FileSystem::TempFolder() {
     return path;
 }
 
-std::string FileSystem::HomeFolder() { return getenv("HOME"); }
+std::string FileSystem::HomeFolder()
+{
+    auto dir = getenv("HOME");
+    return dir == nullptr ? "" : dir;
+}
